video stimulus: partial play works, universal stretch value calculated for all rotations

+++ engine/vision_experiment/stimulation_library.py
@@ -302,8 +302,8 @@ class Stimulations(experiment_control.ExperimentControl):#, screen.ScreenAndKeyb
             files=os.listdir(path)
             files.sort()
             if length>0:
-                length_f=(self.config.SCREEN_EXPECTED_FRAME_RATE*length)
-                offset_f=(self.config.SCREEN_EXPECTED_FRAME_RATE*offset)
+                length_f=int(self.config.SCREEN_EXPECTED_FRAME_RATE*length)
+                offset_f=int(self.config.SCREEN_EXPECTED_FRAME_RATE*offset)
                 files=files[offset_f:offset_f+length_f]
             if len([f for f in files if os.path.splitext(f)[1] not in ['.png', '.bmp', '.jpg']])>0:
                  raise RuntimeError('{0} folder contains non image files, please remove them!'.format(path))
